fix: Extract results step out of main.dart (closes #17)
- Move _ResultsStep widget and tightly coupled helpers into widgets/results_step.dart
- Extracted widgets: ResultsStep, _RecommendationConfidenceBadge, _SpeedComparisonRow, _ResultNarrativeCard, _DarkStatPill, _SpotlightCard
- Keep _SpotlightCard in main.dart for reuse by _UsageStep
- Simplify main.dart by ~400 lines
- Callbacks and state management remain owned by parent RecommendationFlowPage
- Export service tests pass; widget functionality preserved
